// lib/api/sdk/models/create-knowledge-chunk-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Chatbot
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateKnowledgeChunkRequest
 */
export interface CreateKnowledgeChunkRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateKnowledgeChunkRequest
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof CreateKnowledgeChunkRequest
     */
    'departmentId': string;
}